requirements.txt
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-Migrate==2.1.1
flask-restx==0.2.0
Flask-Script==2.0.6
gunicorn==20.0.4
itsdangerous==2.0.1
Jinja2==3.0.1
mysqlclient==2.1.0
pdf2image==1.16.0
Pillow==9.1.0
PyJWT==2.3.0
pyOpenSSL==22.0.0
pytesseract==0.3.9
python-dateutil==2.8.2
python-dotenv==0.15.0
requests==2.25.1
werkzeug==0.16.1